move out of time to its own controller
